+++ src/main/java/org/opensearch/knn/index/query/KNNQueryBuilder.java
@@ -7,6 +7,7 @@ package org.opensearch.knn.index.query;
 
 import java.io.IOException;
 import java.util.Arrays;
+
 import java.util.List;
 import java.util.Objects;
 import lombok.extern.log4j.Log4j2;
@@ -24,6 +25,7 @@ import org.opensearch.index.mapper.NumberFieldMapper;
 import org.opensearch.index.query.AbstractQueryBuilder;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.index.query.QueryShardContext;
+import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.KNNMethodContext;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
@@ -35,6 +37,7 @@ import org.opensearch.knn.plugin.stats.KNNCounter;
 
 import static org.opensearch.knn.common.KNNValidationUtil.validateByteVectorValue;
 import static org.opensearch.knn.index.IndexUtil.isClusterOnOrAfterMinRequiredVersion;
+import static org.opensearch.knn.index.util.KNNEngine.ENGINES_SUPPORTING_RADIAL_SEARCH;
 
 /**
  * Helper class to build the KNN query
@@ -47,6 +50,7 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
     public static final ParseField K_FIELD = new ParseField("k");
     public static final ParseField FILTER_FIELD = new ParseField("filter");
     public static final ParseField IGNORE_UNMAPPED_FIELD = new ParseField("ignore_unmapped");
+    public static final ParseField DISTANCE_FIELD = new ParseField("distance");
     public static final int K_MAX = 10000;
     /**
      * The name for the knn query
@@ -58,11 +62,74 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
     private final String fieldName;
     private final float[] vector;
     private int k = 0;
+    private Float distance = null;
     private QueryBuilder filter;
     private boolean ignoreUnmapped = false;
 
     /**
-     * Constructs a new knn query
+    * Constructs a new query with the given field name and vector
+     *
+     * @param fieldName Name of the field
+     * @param vector    Array of floating points
+    */
+    public KNNQueryBuilder(String fieldName, float[] vector) {
+        if (Strings.isNullOrEmpty(fieldName)) {
+            throw new IllegalArgumentException("[" + NAME + "] requires fieldName");
+        }
+        if (vector == null) {
+            throw new IllegalArgumentException("[" + NAME + "] requires query vector");
+        }
+        if (vector.length == 0) {
+            throw new IllegalArgumentException("[" + NAME + "] query vector is empty");
+        }
+        this.fieldName = fieldName;
+        this.vector = vector;
+    }
+
+    /**
+     * Builder method for k
+     *
+     * @param k K nearest neighbours for the given vector
+     */
+    public KNNQueryBuilder k(int k) {
+        if (k <= 0 || k > K_MAX) {
+            throw new IllegalArgumentException("[" + NAME + "] requires 0 < k <= " + K_MAX);
+        }
+        if (distance != null) {
+            throw new IllegalArgumentException("[" + NAME + "] requires either k or distance must be set");
+        }
+        this.k = k;
+        return this;
+    }
+
+    /**
+     * Builder method for distance
+     *
+     * @param distance the distance threshold for the nearest neighbours
+     */
+    public KNNQueryBuilder distance(Float distance) {
+        if (distance == null || distance < 0) {
+            throw new IllegalArgumentException("[" + NAME + "] requires distance >= 0");
+        }
+        if (k != 0) {
+            throw new IllegalArgumentException("[" + NAME + "] requires either k or distance must be set");
+        }
+        this.distance = distance;
+        return this;
+    }
+
+    /**
+     * Builder method for filter
+     *
+     * @param filter QueryBuilder
+     */
+    public KNNQueryBuilder filter(QueryBuilder filter) {
+        this.filter = filter;
+        return this;
+    }
+
+    /**
+     * Constructs a new query for top k search
      *
      * @param fieldName Name of the filed
      * @param vector    Array of floating points
@@ -94,6 +161,7 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         this.k = k;
         this.filter = filter;
         this.ignoreUnmapped = false;
+        this.distance = null;
     }
 
     public static void initialize(ModelDao modelDao) {
@@ -128,6 +196,9 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
             if (isClusterOnOrAfterMinRequiredVersion("ignore_unmapped")) {
                 ignoreUnmapped = in.readOptionalBoolean();
             }
+            if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
+                distance = in.readOptionalFloat();
+            }
         } catch (IOException ex) {
             throw new RuntimeException("[KNN] Unable to create KNNQueryBuilder", ex);
         }
@@ -137,7 +208,8 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         String fieldName = null;
         List<Object> vector = null;
         float boost = AbstractQueryBuilder.DEFAULT_BOOST;
-        int k = 0;
+        Integer k = null;
+        Float distance = null;
         QueryBuilder filter = null;
         String queryName = null;
         String currentFieldName = null;
@@ -166,6 +238,8 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
                             }
                         } else if (AbstractQueryBuilder.NAME_FIELD.match(currentFieldName, parser.getDeprecationHandler())) {
                             queryName = parser.text();
+                        } else if (DISTANCE_FIELD.match(currentFieldName, parser.getDeprecationHandler())) {
+                            distance = (Float) NumberFieldMapper.NumberType.FLOAT.parse(parser.objectBytes(), false);
                         } else {
                             throw new ParsingException(
                                 parser.getTokenLocation(),
@@ -195,10 +269,21 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
             }
         }
 
-        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(fieldName, ObjectsToFloats(vector), k, filter);
-        knnQueryBuilder.ignoreUnmapped(ignoreUnmapped);
-        knnQueryBuilder.queryName(queryName);
-        knnQueryBuilder.boost(boost);
+        if ((k != null && distance != null) || (k == null && distance == null)) {
+            throw new IllegalArgumentException("[" + NAME + "] requires either k or distance must be set");
+        }
+
+        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(fieldName, ObjectsToFloats(vector)).filter(filter)
+            .ignoreUnmapped(ignoreUnmapped)
+            .boost(boost)
+            .queryName(queryName);
+
+        if (k != null) {
+            knnQueryBuilder.k(k);
+        } else {
+            knnQueryBuilder.distance(distance);
+        }
+
         return knnQueryBuilder;
     }
 
@@ -210,6 +295,9 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         out.writeOptionalNamedWriteable(filter);
         if (isClusterOnOrAfterMinRequiredVersion("ignore_unmapped")) {
             out.writeOptionalBoolean(ignoreUnmapped);
+        }
+        if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
+            out.writeOptionalFloat(distance);
         }
     }
 
@@ -229,6 +317,10 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
 
     public int getK() {
         return this.k;
+    }
+
+    public float getDistance() {
+        return this.distance;
     }
 
     public QueryBuilder getFilter() {
@@ -258,6 +350,9 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         builder.field(K_FIELD.getPreferredName(), k);
         if (filter != null) {
             builder.field(FILTER_FIELD.getPreferredName(), filter);
+        }
+        if (distance != null) {
+            builder.field(DISTANCE_FIELD.getPreferredName(), distance);
         }
         if (ignoreUnmapped) {
             builder.field(IGNORE_UNMAPPED_FIELD.getPreferredName(), ignoreUnmapped);
@@ -298,6 +393,14 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         } else if (knnMethodContext != null) {
             // If the dimension is set but the knnMethodContext is not then the field is using the legacy mapping
             knnEngine = knnMethodContext.getKnnEngine();
+            spaceType = knnMethodContext.getSpaceType();
+        }
+
+        // Currently, k-NN supports distance type radius search.
+        // We need transform distance radius to right type of engine required radius.
+        Float radius = null;
+        if (this.distance != null) {
+            radius = knnEngine.distanceToRadialThreshold(this.distance, spaceType);
         }
 
         if (fieldDimension != vector.length) {
@@ -325,18 +428,40 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         }
 
         String indexName = context.index().getName();
-        KNNQueryFactory.CreateQueryRequest createQueryRequest = KNNQueryFactory.CreateQueryRequest.builder()
-            .knnEngine(knnEngine)
-            .indexName(indexName)
-            .fieldName(this.fieldName)
-            .vector(VectorDataType.FLOAT == vectorDataType ? this.vector : null)
-            .byteVector(VectorDataType.BYTE == vectorDataType ? byteVector : null)
-            .vectorDataType(vectorDataType)
-            .k(this.k)
-            .filter(this.filter)
-            .context(context)
-            .build();
-        return KNNQueryFactory.create(createQueryRequest);
+
+        if (k != 0) {
+            KNNQueryFactory.CreateQueryRequest createQueryRequest = KNNQueryFactory.CreateQueryRequest.builder()
+                .knnEngine(knnEngine)
+                .indexName(indexName)
+                .fieldName(this.fieldName)
+                .vector(VectorDataType.FLOAT == vectorDataType ? this.vector : null)
+                .byteVector(VectorDataType.BYTE == vectorDataType ? byteVector : null)
+                .vectorDataType(vectorDataType)
+                .k(this.k)
+                .filter(this.filter)
+                .context(context)
+                .build();
+            return KNNQueryFactory.create(createQueryRequest);
+        }
+        if (radius != null) {
+            if (!ENGINES_SUPPORTING_RADIAL_SEARCH.contains(knnEngine)) {
+                throw new UnsupportedOperationException(String.format("Engine [%s] does not support radial search", knnEngine));
+            }
+            RNNQueryFactory.CreateQueryRequest createQueryRequest = RNNQueryFactory.CreateQueryRequest.builder()
+                .knnEngine(knnEngine)
+                .indexName(indexName)
+                .fieldName(this.fieldName)
+                .vector(VectorDataType.FLOAT == vectorDataType ? this.vector : null)
+                .byteVector(VectorDataType.BYTE == vectorDataType ? byteVector : null)
+                .vectorDataType(vectorDataType)
+                .radius(radius)
+                .filter(this.filter)
+                .context(context)
+                .radius(radius)
+                .build();
+            return RNNQueryFactory.create(createQueryRequest);
+        }
+        throw new IllegalArgumentException("[" + NAME + "] requires either k or distance must be set");
     }
 
     private ModelMetadata getModelMetadataForField(KNNVectorFieldMapper.KNNVectorFieldType knnVectorField) {