+++ src/main/java/org/opensearch/knn/index/query/KNNQueryFactory.java
@@ -5,11 +5,6 @@
 
 package org.opensearch.knn.index.query;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.Setter;
 import lombok.extern.log4j.Log4j2;
 import org.apache.lucene.search.KnnByteVectorQuery;
 import org.apache.lucene.search.KnnFloatVectorQuery;
@@ -17,16 +12,11 @@ import org.apache.lucene.search.Query;
 import org.apache.lucene.search.join.BitSetProducer;
 import org.apache.lucene.search.join.DiversifyingChildrenByteKnnVectorQuery;
 import org.apache.lucene.search.join.DiversifyingChildrenFloatKnnVectorQuery;
-import org.apache.lucene.search.join.ToChildBlockJoinQuery;
-import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.index.query.QueryShardContext;
-import org.opensearch.index.search.NestedHelper;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.util.KNNEngine;
 
-import java.io.IOException;
 import java.util.Locale;
-import java.util.Optional;
 
 import static org.opensearch.knn.common.KNNConstants.VECTOR_DATA_TYPE_FIELD;
 import static org.opensearch.knn.index.VectorDataType.SUPPORTED_VECTOR_DATA_TYPES;
@@ -35,7 +25,7 @@ import static org.opensearch.knn.index.VectorDataType.SUPPORTED_VECTOR_DATA_TYPE
  * Creates the Lucene k-NN queries
  */
 @Log4j2
-public class KNNQueryFactory {
+public class KNNQueryFactory extends BaseQueryFactory {
 
     /**
      * Creates a Lucene query for a particular engine.
@@ -82,7 +72,12 @@ public class KNNQueryFactory {
         final VectorDataType vectorDataType = createQueryRequest.getVectorDataType();
         final Query filterQuery = getFilterQuery(createQueryRequest);
 
-        BitSetProducer parentFilter = createQueryRequest.context == null ? null : createQueryRequest.context.getParentFilter();
+        BitSetProducer parentFilter = null;
+        if (createQueryRequest.getContext().isPresent()) {
+            QueryShardContext context = createQueryRequest.getContext().get();
+            parentFilter = context.getParentFilter();
+        }
+
         if (KNNEngine.getEnginesThatCreateCustomSegmentFiles().contains(createQueryRequest.getKnnEngine())) {
             if (filterQuery != null && KNNEngine.getEnginesThatSupportsFilters().contains(createQueryRequest.getKnnEngine())) {
                 log.debug("Creating custom k-NN query with filters for index: {}, field: {} , k: {}", indexName, fieldName, k);
@@ -93,19 +88,21 @@ public class KNNQueryFactory {
         }
 
         log.debug(String.format("Creating Lucene k-NN query for index: %s \"\", field: %s \"\", k: %d", indexName, fieldName, k));
-        if (VectorDataType.BYTE == vectorDataType) {
-            return getKnnByteVectorQuery(fieldName, byteVector, k, filterQuery, parentFilter);
-        } else if (VectorDataType.FLOAT == vectorDataType) {
-            return getKnnFloatVectorQuery(fieldName, vector, k, filterQuery, parentFilter);
-        } else {
-            throw new IllegalArgumentException(
-                String.format(
-                    Locale.ROOT,
-                    "Invalid value provided for [%s] field. Supported values are [%s]",
-                    VECTOR_DATA_TYPE_FIELD,
-                    SUPPORTED_VECTOR_DATA_TYPES
-                )
-            );
+        switch (vectorDataType) {
+            case BYTE:
+                return getKnnByteVectorQuery(fieldName, byteVector, k, filterQuery, parentFilter);
+            case FLOAT:
+                return getKnnFloatVectorQuery(fieldName, vector, k, filterQuery, parentFilter);
+            default:
+                throw new IllegalArgumentException(
+                    String.format(
+                        Locale.ROOT,
+                        "Invalid value provided for [%s] field. Supported values are [%s], but got: %s",
+                        VECTOR_DATA_TYPE_FIELD,
+                        SUPPORTED_VECTOR_DATA_TYPES,
+                        vectorDataType
+                    )
+                );
         }
     }
 
@@ -142,79 +139,6 @@ public class KNNQueryFactory {
             return new KnnFloatVectorQuery(fieldName, floatVector, k, filterQuery);
         } else {
             return new DiversifyingChildrenFloatKnnVectorQuery(fieldName, floatVector, filterQuery, k, parentFilter);
-        }
-    }
-
-    private static Query getFilterQuery(CreateQueryRequest createQueryRequest) {
-        if (createQueryRequest.getFilter().isPresent()) {
-            final QueryShardContext queryShardContext = createQueryRequest.getContext()
-                .orElseThrow(() -> new RuntimeException("Shard context cannot be null"));
-            log.debug(
-                String.format(
-                    "Creating k-NN query with filter for index [%s], field [%s] and k [%d]",
-                    createQueryRequest.getIndexName(),
-                    createQueryRequest.fieldName,
-                    createQueryRequest.k
-                )
-            );
-            final Query filterQuery;
-            try {
-                filterQuery = createQueryRequest.getFilter().get().toQuery(queryShardContext);
-            } catch (IOException e) {
-                throw new RuntimeException("Cannot create knn query with filter", e);
-            }
-            // If k-NN Field is nested field then parentFilter will not be null. This parentFilter is set by the
-            // Opensearch core. Ref PR: https://github.com/opensearch-project/OpenSearch/pull/10246
-            if (queryShardContext.getParentFilter() != null) {
-                // if the filter is also a nested query clause then we should just return the same query without
-                // considering it to join with the parent documents.
-                if (new NestedHelper(queryShardContext.getMapperService()).mightMatchNestedDocs(filterQuery)) {
-                    return filterQuery;
-                }
-                // This condition will be hit when filters are getting applied on the top level fields and k-nn
-                // query field is a nested field. In this case we need to wrap the filter query with
-                // ToChildBlockJoinQuery to ensure parent documents which will be retrieved from filters can be
-                // joined with the child documents containing vector field.
-                return new ToChildBlockJoinQuery(filterQuery, queryShardContext.getParentFilter());
-            }
-            return filterQuery;
-        }
-        return null;
-    }
-
-    /**
-     * DTO object to hold data required to create a Query instance.
-     */
-    @AllArgsConstructor
-    @Builder
-    @Setter
-    static class CreateQueryRequest {
-        @Getter
-        @NonNull
-        private KNNEngine knnEngine;
-        @Getter
-        @NonNull
-        private String indexName;
-        @Getter
-        private String fieldName;
-        @Getter
-        private float[] vector;
-        @Getter
-        private byte[] byteVector;
-        @Getter
-        private VectorDataType vectorDataType;
-        @Getter
-        private int k;
-        private QueryBuilder filter;
-        // can be null in cases filter not passed with the knn query
-        private QueryShardContext context;
-
-        public Optional<QueryBuilder> getFilter() {
-            return Optional.ofNullable(filter);
-        }
-
-        public Optional<QueryShardContext> getContext() {
-            return Optional.ofNullable(context);
         }
     }
 }