+++ src/test/java/org/opensearch/knn/index/query/KNNQueryBuilderTests.java
@@ -6,6 +6,7 @@
 package org.opensearch.knn.index.query;
 
 import com.google.common.collect.ImmutableMap;
+import org.apache.lucene.search.FloatVectorSimilarityQuery;
 import java.util.Locale;
 import org.apache.lucene.search.KnnFloatVectorQuery;
 import org.apache.lucene.search.MatchNoDocsQuery;
@@ -41,8 +42,10 @@ import org.opensearch.plugins.SearchPlugin;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static org.hamcrest.Matchers.instanceOf;
 import static org.mockito.Mockito.anyString;
@@ -50,11 +53,13 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.opensearch.knn.common.KNNConstants.METHOD_HNSW;
 import static org.opensearch.knn.index.KNNClusterTestUtils.mockClusterService;
+import static org.opensearch.knn.index.util.KNNEngine.ENGINES_SUPPORTING_RADIAL_SEARCH;
 
 public class KNNQueryBuilderTests extends KNNTestCase {
 
     private static final String FIELD_NAME = "myvector";
     private static final int K = 1;
+    private static final Float DISTANCE = 1.0f;
     private static final TermQueryBuilder TERM_QUERY = QueryBuilders.termQuery("field", "value");
     private static final float[] QUERY_VECTOR = new float[] { 1.0f, 2.0f, 3.0f, 4.0f };
 
@@ -77,6 +82,20 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector, KNNQueryBuilder.K_MAX + K));
     }
 
+    public void testInvalidDistance() {
+        float[] queryVector = { 1.0f, 1.0f };
+
+        /**
+         * -ve distance
+         */
+        expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector).distance(-1.0f));
+
+        /**
+         * null distance
+         */
+        expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector).distance(null));
+    }
+
     public void testEmptyVector() {
         /**
          * null query vector
@@ -89,6 +108,18 @@ public class KNNQueryBuilderTests extends KNNTestCase {
          */
         float[] queryVector1 = {};
         expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector1, K));
+
+        /**
+         * null query vector with distance
+         */
+        float[] queryVector2 = null;
+        expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector2).distance(DISTANCE));
+
+        /**
+         * empty query vector with distance
+         */
+        float[] queryVector3 = {};
+        expectThrows(IllegalArgumentException.class, () -> new KNNQueryBuilder(FIELD_NAME, queryVector3).distance(DISTANCE));
     }
 
     public void testFromXContent() throws Exception {
@@ -107,7 +138,23 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         assertEquals(knnQueryBuilder, actualBuilder);
     }
 
-    public void testFromXContent_WithFilter() throws Exception {
+    public void testFromXContent_whenDoRadiusSearch_thenSucceed() throws Exception {
+        float[] queryVector = { 1.0f, 2.0f, 3.0f, 4.0f };
+        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, queryVector).distance(DISTANCE);
+        XContentBuilder builder = XContentFactory.jsonBuilder();
+        builder.startObject();
+        builder.startObject(knnQueryBuilder.fieldName());
+        builder.field(KNNQueryBuilder.VECTOR_FIELD.getPreferredName(), knnQueryBuilder.vector());
+        builder.field(KNNQueryBuilder.DISTANCE_FIELD.getPreferredName(), knnQueryBuilder.getDistance());
+        builder.endObject();
+        builder.endObject();
+        XContentParser contentParser = createParser(builder);
+        contentParser.nextToken();
+        KNNQueryBuilder actualBuilder = KNNQueryBuilder.fromXContent(contentParser);
+        assertEquals(knnQueryBuilder, actualBuilder);
+    }
+
+    public void testFromXContent_withFilter() throws Exception {
         final ClusterService clusterService = mockClusterService(Version.CURRENT);
 
         final KNNClusterUtil knnClusterUtil = KNNClusterUtil.instance();
@@ -129,7 +176,29 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         assertEquals(knnQueryBuilder, actualBuilder);
     }
 
-    public void testFromXContent_invalidQueryVectorType() throws Exception {
+    public void testFromXContent_wenDoRadiusSearch_whenFilter_thenSucceed() throws Exception {
+        final ClusterService clusterService = mockClusterService(Version.CURRENT);
+
+        final KNNClusterUtil knnClusterUtil = KNNClusterUtil.instance();
+        knnClusterUtil.initialize(clusterService);
+
+        float[] queryVector = { 1.0f, 2.0f, 3.0f, 4.0f };
+        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, queryVector).distance(DISTANCE).filter(TERM_QUERY);
+        XContentBuilder builder = XContentFactory.jsonBuilder();
+        builder.startObject();
+        builder.startObject(knnQueryBuilder.fieldName());
+        builder.field(KNNQueryBuilder.VECTOR_FIELD.getPreferredName(), knnQueryBuilder.vector());
+        builder.field(KNNQueryBuilder.DISTANCE_FIELD.getPreferredName(), knnQueryBuilder.getDistance());
+        builder.field(KNNQueryBuilder.FILTER_FIELD.getPreferredName(), knnQueryBuilder.getFilter());
+        builder.endObject();
+        builder.endObject();
+        XContentParser contentParser = createParser(builder);
+        contentParser.nextToken();
+        KNNQueryBuilder actualBuilder = KNNQueryBuilder.fromXContent(contentParser);
+        assertEquals(knnQueryBuilder, actualBuilder);
+    }
+
+    public void testFromXContent_InvalidQueryVectorType() throws Exception {
         final ClusterService clusterService = mockClusterService(Version.CURRENT);
 
         final KNNClusterUtil knnClusterUtil = KNNClusterUtil.instance();
@@ -146,6 +215,34 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         builder.startObject(FIELD_NAME);
         builder.field(KNNQueryBuilder.VECTOR_FIELD.getPreferredName(), invalidTypeQueryVector);
         builder.field(KNNQueryBuilder.K_FIELD.getPreferredName(), K);
+        builder.endObject();
+        builder.endObject();
+        XContentParser contentParser = createParser(builder);
+        contentParser.nextToken();
+        IllegalArgumentException exception = expectThrows(
+            IllegalArgumentException.class,
+            () -> KNNQueryBuilder.fromXContent(contentParser)
+        );
+        assertTrue(exception.getMessage().contains("[knn] field 'vector' requires to be an array of numbers"));
+    }
+
+    public void testFromXContent_whenDoRadiusSearch_whenInputInvalidQueryVectorType_thenException() throws Exception {
+        final ClusterService clusterService = mockClusterService(Version.CURRENT);
+
+        final KNNClusterUtil knnClusterUtil = KNNClusterUtil.instance();
+        knnClusterUtil.initialize(clusterService);
+
+        List<Object> invalidTypeQueryVector = new ArrayList<>();
+        invalidTypeQueryVector.add(1.5);
+        invalidTypeQueryVector.add(2.5);
+        invalidTypeQueryVector.add("a");
+        invalidTypeQueryVector.add(null);
+
+        XContentBuilder builder = XContentFactory.jsonBuilder();
+        builder.startObject();
+        builder.startObject(FIELD_NAME);
+        builder.field(KNNQueryBuilder.VECTOR_FIELD.getPreferredName(), invalidTypeQueryVector);
+        builder.field(KNNQueryBuilder.DISTANCE_FIELD.getPreferredName(), DISTANCE);
         builder.endObject();
         builder.endObject();
         XContentParser contentParser = createParser(builder);
@@ -231,6 +328,23 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         assertEquals(knnQueryBuilder.vector(), query.getQueryVector());
     }
 
+    public void testDoToQuery_whenNormal_whenDoRadiusSearch_whenSucceed() {
+        float[] queryVector = { 1.0f, 2.0f, 3.0f, 4.0f };
+        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, queryVector).distance(DISTANCE);
+        Index dummyIndex = new Index("dummy", "dummy");
+        QueryShardContext mockQueryShardContext = mock(QueryShardContext.class);
+        KNNVectorFieldMapper.KNNVectorFieldType mockKNNVectorField = mock(KNNVectorFieldMapper.KNNVectorFieldType.class);
+        when(mockQueryShardContext.index()).thenReturn(dummyIndex);
+        when(mockKNNVectorField.getDimension()).thenReturn(4);
+        when(mockKNNVectorField.getVectorDataType()).thenReturn(VectorDataType.FLOAT);
+        when(mockQueryShardContext.fieldMapper(anyString())).thenReturn(mockKNNVectorField);
+        MethodComponentContext methodComponentContext = new MethodComponentContext(METHOD_HNSW, ImmutableMap.of());
+        KNNMethodContext knnMethodContext = new KNNMethodContext(KNNEngine.LUCENE, SpaceType.L2, methodComponentContext);
+        when(mockKNNVectorField.getKnnMethodContext()).thenReturn(knnMethodContext);
+        FloatVectorSimilarityQuery query = (FloatVectorSimilarityQuery) knnQueryBuilder.doToQuery(mockQueryShardContext);
+        assertTrue(query.toString().contains("resultSimilarity=" + KNNEngine.LUCENE.distanceToRadialThreshold(DISTANCE, SpaceType.L2)));
+    }
+
     public void testDoToQuery_KnnQueryWithFilter() throws Exception {
         float[] queryVector = { 1.0f, 2.0f, 3.0f, 4.0f };
         KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, queryVector, K, TERM_QUERY);
@@ -248,6 +362,24 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         Query query = knnQueryBuilder.doToQuery(mockQueryShardContext);
         assertNotNull(query);
         assertTrue(query.getClass().isAssignableFrom(KnnFloatVectorQuery.class));
+    }
+
+    public void testDoToQuery_whenDoRadiusSearch_whenFilter_thenSucceed() {
+        float[] queryVector = { 1.0f, 2.0f, 3.0f, 4.0f };
+        KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, queryVector).distance(DISTANCE).filter(TERM_QUERY);
+        Index dummyIndex = new Index("dummy", "dummy");
+        QueryShardContext mockQueryShardContext = mock(QueryShardContext.class);
+        KNNVectorFieldMapper.KNNVectorFieldType mockKNNVectorField = mock(KNNVectorFieldMapper.KNNVectorFieldType.class);
+        when(mockQueryShardContext.index()).thenReturn(dummyIndex);
+        when(mockKNNVectorField.getDimension()).thenReturn(4);
+        when(mockKNNVectorField.getVectorDataType()).thenReturn(VectorDataType.FLOAT);
+        MethodComponentContext methodComponentContext = new MethodComponentContext(METHOD_HNSW, ImmutableMap.of());
+        KNNMethodContext knnMethodContext = new KNNMethodContext(KNNEngine.LUCENE, SpaceType.L2, methodComponentContext);
+        when(mockKNNVectorField.getKnnMethodContext()).thenReturn(knnMethodContext);
+        when(mockQueryShardContext.fieldMapper(anyString())).thenReturn(mockKNNVectorField);
+        Query query = knnQueryBuilder.doToQuery(mockQueryShardContext);
+        assertNotNull(query);
+        assertTrue(query.getClass().isAssignableFrom(FloatVectorSimilarityQuery.class));
     }
 
     public void testDoToQuery_WhenknnQueryWithFilterAndFaissEngine_thenSuccess() {
@@ -383,17 +515,18 @@ public class KNNQueryBuilderTests extends KNNTestCase {
     }
 
     public void testSerialization() throws Exception {
-        assertSerialization(Version.CURRENT, Optional.empty());
+        assertSerialization(Version.CURRENT, Optional.empty(), K, null);
+        assertSerialization(Version.CURRENT, Optional.of(TERM_QUERY), K, null);
+        assertSerialization(Version.V_2_3_0, Optional.empty(), K, null);
 
-        assertSerialization(Version.CURRENT, Optional.of(TERM_QUERY));
-
-        assertSerialization(Version.V_2_3_0, Optional.empty());
+        // For radius search
+        assertSerialization(Version.CURRENT, Optional.empty(), null, DISTANCE);
+        assertSerialization(Version.CURRENT, Optional.of(TERM_QUERY), null, DISTANCE);
     }
 
-    private void assertSerialization(final Version version, final Optional<QueryBuilder> queryBuilderOptional) throws Exception {
-        final KNNQueryBuilder knnQueryBuilder = queryBuilderOptional.isPresent()
-            ? new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR, K, queryBuilderOptional.get())
-            : new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR, K);
+    private void assertSerialization(final Version version, final Optional<QueryBuilder> queryBuilderOptional, Integer k, Float distance)
+        throws Exception {
+        final KNNQueryBuilder knnQueryBuilder = getKnnQueryBuilder(queryBuilderOptional, k, distance);
 
         final ClusterService clusterService = mockClusterService(version);
 
@@ -412,7 +545,11 @@ public class KNNQueryBuilderTests extends KNNTestCase {
                 final KNNQueryBuilder deserializedKnnQueryBuilder = (KNNQueryBuilder) deserializedQuery;
                 assertEquals(FIELD_NAME, deserializedKnnQueryBuilder.fieldName());
                 assertArrayEquals(QUERY_VECTOR, (float[]) deserializedKnnQueryBuilder.vector(), 0.0f);
-                assertEquals(K, deserializedKnnQueryBuilder.getK());
+                if (k != null) {
+                    assertEquals(k.intValue(), deserializedKnnQueryBuilder.getK());
+                } else {
+                    assertEquals(distance.floatValue(), deserializedKnnQueryBuilder.getDistance(), 0.0f);
+                }
                 if (queryBuilderOptional.isPresent()) {
                     assertNotNull(deserializedKnnQueryBuilder.getFilter());
                     assertEquals(queryBuilderOptional.get(), deserializedKnnQueryBuilder.getFilter());
@@ -421,6 +558,22 @@ public class KNNQueryBuilderTests extends KNNTestCase {
                 }
             }
         }
+    }
+
+    private static KNNQueryBuilder getKnnQueryBuilder(Optional<QueryBuilder> queryBuilderOptional, Integer k, Float distance) {
+        final KNNQueryBuilder knnQueryBuilder;
+        if (k != null) {
+            knnQueryBuilder = queryBuilderOptional.isPresent()
+                ? new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR, k, queryBuilderOptional.get())
+                : new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR, k);
+        } else if (distance != null) {
+            knnQueryBuilder = queryBuilderOptional.isPresent()
+                ? new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR).distance(distance).filter(queryBuilderOptional.get())
+                : new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR).distance(distance);
+        } else {
+            throw new IllegalArgumentException("Either k or distance must be provided");
+        }
+        return knnQueryBuilder;
     }
 
     public void testIgnoreUnmapped() throws IOException {
@@ -433,5 +586,28 @@ public class KNNQueryBuilderTests extends KNNTestCase {
         assertThat(query, instanceOf(MatchNoDocsQuery.class));
         knnQueryBuilder.ignoreUnmapped(false);
         expectThrows(IllegalArgumentException.class, () -> knnQueryBuilder.doToQuery(mock(QueryShardContext.class)));
+    }
+
+    public void testRadialSearch_whenUnsupportedEngine_thenThrowException() {
+        List<KNNEngine> unsupportedEngines = Arrays.stream(KNNEngine.values())
+            .filter(knnEngine -> !ENGINES_SUPPORTING_RADIAL_SEARCH.contains(knnEngine))
+            .collect(Collectors.toList());
+        for (KNNEngine knnEngine : unsupportedEngines) {
+            KNNMethodContext knnMethodContext = new KNNMethodContext(
+                knnEngine,
+                SpaceType.L2,
+                new MethodComponentContext(METHOD_HNSW, ImmutableMap.of())
+            );
+            KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(FIELD_NAME, QUERY_VECTOR).distance(DISTANCE);
+            KNNVectorFieldMapper.KNNVectorFieldType mockKNNVectorField = mock(KNNVectorFieldMapper.KNNVectorFieldType.class);
+            QueryShardContext mockQueryShardContext = mock(QueryShardContext.class);
+            Index dummyIndex = new Index("dummy", "dummy");
+            when(mockKNNVectorField.getKnnMethodContext()).thenReturn(knnMethodContext);
+            when(mockQueryShardContext.index()).thenReturn(dummyIndex);
+            when(mockKNNVectorField.getDimension()).thenReturn(4);
+            when(mockQueryShardContext.fieldMapper(anyString())).thenReturn(mockKNNVectorField);
+
+            expectThrows(UnsupportedOperationException.class, () -> knnQueryBuilder.doToQuery(mockQueryShardContext));
+        }
     }
 }