+++ src/test/java/org/opensearch/knn/index/util/AbstractKNNLibraryTests.java
@@ -128,6 +128,11 @@ public class AbstractKNNLibraryTests extends KNNTestCase {
         }
 
         @Override
+        public Float distanceToRadialThreshold(Float distance, SpaceType spaceType) {
+            return 0f;
+        }
+
+        @Override
         public int estimateOverheadInKB(KNNMethodContext knnMethodContext, int dimension) {
             return 0;
         }